third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/iframe-inside-csp-expected.txt

CONSOLE MESSAGE: line 2: PASS (1/2): Script can execute
CONSOLE MESSAGE: PASS (2/2): Eval works
